Refactor pop/push/shift/unshift methods with trait
